requirements.txt
#dvc==3.51.2
#dvc-gdrive==3.0.1
fastapi==0.111.1
imbalanced-learn==0.12.3
joblib==1.4.2
mlflow==2.14.3
mkdocs==1.6.0
mkdocs-material==9.5.31
numpy
pandas
pytest==8.3.2
PyYAML==6.0.1
scikit-learn
setuptools==70.3.0
uvicorn==0.30.1
matplotlib
# evidently @ git+https://github.com/evidentlyai/evidently.git